pyproject.toml
[project]
name = "devmachine"
version = "0.0.1"
requires-python = ">=3.14"
dependencies = [
    "fabric>=3.0,<4",
    "pyyaml",
    "mypy>=1.18.1",
    "ruff>=0.13.0",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.ruff]
lint.select = ["E4", "E7", "E9", "F", "I"]

[tool.uv]
package = false

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.python.org/simple"
